Remove skip_callback, set_callback
The main reason is to remove `set_callback ...., :update, .....`, because
it registers `:update` filter in the `create` callback chains.
That means that on the `create` event, `update` method is called,
which then causes to send 2x 'save' event, and all
hooks that are listening on that event are triggered twice.

Other than that it doesn't make any difference if these lines are in
the code. Tests are still passing, and it doesn't call any
callbacks when it runs migration.